Generate /etc/hosts during rebuild
We'd like to avoid maintaining the 'hosts' file, so we download a fresh copy
from someonewhocares.org during the rebuild.

This is a security concern though: we don't want to allow arbitrary redirects to
get added to our /etc/hosts! To avoid this, we strip downloaded file down to
just a list of domains, and we stick 127.0.0.1 in front of them ourselves. This
way, the worst thing a malicious entry could do is cause a domain to get
blocked; it can't redirect us to somewhere else.

Whilst we're at it, we might as well add some extra facebook domains, since I've
noticed that some of their Javascript is getting through.