Make Uni monitor "primary", so xmobar and xmonad work together
Without this, Xmonad seems to leave a space for xmobar on the external monitor,
but actually displays it on the internal monitor (behind any other windows).

Here we add the '--primary' option for our plug/unplug scripts, which causes
xmobar to appear in the space that xmonad gives it on the external monitor. When
unplugged, the space appears on the internal monitor and xmobar appears as
expected.