Use qterminal rather than foot
Foot is fast and lightweight, but it barfs on certain inputs. For example,
typing a rather long command (say, a loop), then trying to go back to edit some
earlier part, Foot will just mangle the display, so it's impossible to tell
whereabouts in the command we are.

I tried a few alternative terminals, and found the following:
 - KingsCross (AKA kgx) came with Phosh, but it takes a HUGE amount of
   CPU (e.g. ~70% for htop refreshing every 2 secs), and has awful
   latency (literally having to wait several seconds for it to show the command
   I've been typing out)
 - Konsole: this depends on a whole bunch of KDE libraries, which takes up more
   space than I'm comfortable with (until I start using KDE again...)
 - WezTerm: This flat out doesn't work, seemingly due to no EGL. Funnily enough
   it gives an error mesage like "failed with message: SUCCESS"...

In comparison, qterminal seems fast, reasonably low resource-usage, and is
mostly self-contained (no huge dependency tree). It also seems to deal fine with
plugging external monitors in and out, whilst Foot would occasionally make the
font ridiculously huge for no reason.